Add backend database structure
The backend should be device agnostic, but currently the only device
tested with it is the Polar Vantage M. The table columns should be
applicable for other sport watches as well.

The first plugin to use the backend is likely the aforementioned device
type, but it should be simple to use a different brand. This repository
will not contain the different device plugins, and rather be completely
ignorant of who will use it.